+++ common/src/main/java/com/zigythebird/playeranim/animation/keyframe/event/SoundKeyframeEvent.java
@@ -11,12 +11,12 @@ import com.zigythebird.playeranim.animation.keyframe.event.data.SoundKeyframeDat
 import net.minecraft.client.player.AbstractClientPlayer;
 
 /**
- * The {@link KeyFrameEvent} specific to the {@link AnimationController#soundKeyframeHandler}
+ * The {@link KeyFrameEvent} specific to the {@link AnimationController#setSoundKeyframeHandler(AnimationController.SoundKeyframeHandler)}
  * <p>
  * Called when a sound instruction keyframe is encountered
  */
 public class SoundKeyframeEvent extends KeyFrameEvent<SoundKeyframeData> {
-	public SoundKeyframeEvent(AbstractClientPlayer player, double animationTick, AnimationController controller,
+	public SoundKeyframeEvent(AbstractClientPlayer player, float animationTick, AnimationController controller,
 							  SoundKeyframeData keyFrameData, AnimationData animationData) {
 		super(player, animationTick, controller, keyFrameData, animationData);
 	}
